make asc/desc in sort optional and let visitor decide on default
This should make #10 work as described. It still keeps + and - as a reserved words, even when using the second
param as sort order specifier you won't be able to have field names that start with either plus or minus.

For the MongoODM visitor it has been decided to use ``asc`` as the default case, I believe that is sensible because
it is also mongodbs default.

A visitor who wishes to implement the ``sort(<field>, <sortspec>)`` case will still need to unpack the field after getting
it from the AST.
You will be getting two different fields: ``getField() -> array(<field>)`` and ``getField() -> array(<sortspec>)``.